Combine work into a single mojo #43
