+++ aemanalyser-maven-plugin/src/main/java/com/adobe/aem/analyser/mojos/AnalyseMojo.java
@@ -11,149 +11,14 @@
 */
 package com.adobe.aem.analyser.mojos;
 
-import java.io.IOException;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Properties;
-import java.util.Set;
-
-import com.adobe.aem.analyser.AemAnalyser;
-import com.adobe.aem.analyser.AemAnalyserResult;
-
-import org.apache.maven.artifact.handler.manager.ArtifactHandlerManager;
-import org.apache.maven.artifact.resolver.ArtifactResolver;
-import org.apache.maven.execution.MavenSession;
-import org.apache.maven.plugin.AbstractMojo;
-import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.MojoFailureException;
-import org.apache.maven.plugins.annotations.Component;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
-import org.apache.maven.plugins.annotations.Parameter;
-import org.apache.maven.project.MavenProject;
-import org.apache.sling.feature.ArtifactId;
-import org.apache.sling.feature.Feature;
-import org.apache.sling.feature.builder.ArtifactProvider;
-import org.apache.sling.feature.io.artifacts.ArtifactManager;
-import org.apache.sling.feature.io.artifacts.ArtifactManagerConfig;
+import org.apache.maven.plugins.annotations.ResolutionScope;
 
-@Mojo(name = "analyse", defaultPhase = LifecyclePhase.TEST)
-public class AnalyseMojo extends AbstractMojo {
+@Mojo(name = "analyse", 
+    defaultPhase = LifecyclePhase.TEST,
+    requiresDependencyResolution = ResolutionScope.COMPILE)
+public class AnalyseMojo extends AemAnalyseMojo {
+    // Subclassed to be used for the aem-analyse packaging
 
-    @Parameter(defaultValue = AemAnalyser.DEFAULT_TASKS,
-        property = "includeTasks")
-    List<String> includeTasks;
-
-    @Parameter
-    Map<String, Properties> taskConfiguration;
-
-    @Parameter(defaultValue = MojoUtils.DEFAULT_SKIP_ENV_VAR, property = MojoUtils.PROPERTY_SKIP_VAR)
-    String skipEnvVarName;
-
-    @Parameter(defaultValue = "true", property = "failon.analyser.errors")
-    private boolean failOnAnalyserErrors;
-
-    @Parameter(property = "project", readonly = true, required = true)
-    protected MavenProject project;
-
-    @Component
-    protected ArtifactHandlerManager artifactHandlerManager;
-
-    @Component
-    protected ArtifactResolver artifactResolver;
-
-    @Parameter(property = "session", readonly = true, required = true)
-    protected MavenSession mavenSession;
-
-
-    ArtifactProvider getArtifactProvider(final ArtifactProvider localProvider) {
-        return new ArtifactProvider() {
-
-            @Override
-            public URL provide(final ArtifactId id) {
-                URL url = localProvider != null ? localProvider.provide(id) : null;
-                if (url != null) {
-                    return url;
-                }
-                try {
-                    return MojoUtils.getOrResolveArtifact(project, mavenSession, artifactHandlerManager, artifactResolver, id).getFile().toURI().toURL();
-                } catch (final MalformedURLException e) {
-                    getLog().debug("Malformed url " + e.getMessage(), e);
-                    // ignore
-                    return null;
-                }
-            }
-        };
-    }
-
-    Map<String, Map<String, String>> getTaskConfigurations() {
-        Map<String, Map<String, String>> config = new HashMap<>();
-        if (this.taskConfiguration != null) {
-            for(final Map.Entry<String, Properties> entry : this.taskConfiguration.entrySet()) {
-                final Map<String, String> m = new HashMap<>();
-
-                entry.getValue().stringPropertyNames().forEach(n -> m.put(n, entry.getValue().getProperty(n)));
-                config.put(entry.getKey(), m);
-            }
-        }
-
-        return config;
-    }
-
-    Set<String> getIncludedTasks() {
-        return new LinkedHashSet<>(this.includeTasks);
-    }
-
-    ArtifactProvider getLocalArtifactProvider() throws IOException {
-        ArtifactManagerConfig amcfg = new ArtifactManagerConfig();
-        amcfg.setRepositoryUrls(new String[] { MojoUtils.getConversionOutputDir(project).toURI().toURL().toString() });
-
-        return ArtifactManager.getArtifactManager(amcfg);
-    }
-    
-    @Override
-    public void execute() throws MojoExecutionException, MojoFailureException {
-        if (MojoUtils.skipRun(skipEnvVarName)) {
-            getLog().info("Skipping AEM analyser plugin as variable " + skipEnvVarName + " is set.");
-            return;
-        }
-
-        boolean hasErrors = false;
-        try {
-            final AemAnalyser analyser = new AemAnalyser();
-            analyser.setArtifactProvider(getArtifactProvider(getLocalArtifactProvider()));
-            analyser.setIncludedTasks(this.getIncludedTasks());
-            analyser.setTaskConfigurations(this.getTaskConfigurations());
-
-            getLog().debug("Retrieving Feature files...");
-            final Collection<Feature> features = MojoUtils.getAggregates(project);
-
-            final AemAnalyserResult result = analyser.analyse(features);
-            
-            for(final String msg : result.getWarnings()) {
-                getLog().warn(msg);
-            }
-            for(final String msg : result.getErrors()) {
-                getLog().error(msg);
-            }
-            hasErrors = result.hasErrors();
-
-        } catch ( final Exception e) {
-            throw new MojoExecutionException("A fatal error occurred while analysing the features, see error cause:",
-                    e);
-        }
-
-        if (hasErrors) {
-            if ( failOnAnalyserErrors ) {
-                throw new MojoFailureException(
-                    "One or more feature analyser(s) detected feature error(s), please read the plugin log for more details");
-            }
-            getLog().warn("Errors found during analyser run, but this plugin is configured to ignore errors and continue the build!");
-        }
-    }
 }
