+++ aemanalyser-maven-plugin/src/main/resources/META-INF/m2e/lifecycle-mapping-metadata.xml
@@ -3,9 +3,8 @@
     <pluginExecution>
       <pluginExecutionFilter>
         <goals>
-          <goal>convert</goal>
-          <goal>aggregate</goal>
           <goal>analyse</goal>
+          <goal>project-analyse</goal>
         </goals>
       </pluginExecutionFilter>
       <action>
