+++ aemanalyser-maven-plugin/src/main/resources/META-INF/plexus/components.xml
@@ -20,8 +20,6 @@
           <lifecycle>
             <id>default</id>
             <phases>
-              <generate-resources>com.adobe.aem:aemanalyser-maven-plugin:convert</generate-resources>
-              <generate-test-resources>com.adobe.aem:aemanalyser-maven-plugin:aggregate</generate-test-resources>
               <test>com.adobe.aem:aemanalyser-maven-plugin:analyse</test>
             </phases>
           </lifecycle>
